Add initialisation routine and other defines to make code more portable
Experimenting with a new way to access PIO registers which allows the PORT to be set for specific pins by header definitions as well as the PINs. So if a pin allocation were to change to a completely different port, the initialisation code for that pin wouldn't have to be rewritten, just the PORT define in the header file would be updated. Additionally, update ADC channel defines in adc_interface.