Improve rotateToHeading
Now it will always take the shortest path to a heading, ie if current heading is -120 and the desired heading is 120, it will go from -120 to -180 then from 180 down to 120 rather than from -120 to 120.
It is also now much more accurate at reaching its set heading. Checks delta yaw on the gyro to see when it has almost finished rotating before exiting with a success status.