Add custom camera read functions to the TWI driver. Got camera initialising.
A camera read operation works in two operations now. The first will send a single byte to the camera in write mode telling it what register to read from. The second reads a single byte (and assumes no internal registers) from the camera. Because the internal register address had already been set in the previous operation, the camera simply replies in the second operation with the data stored in that byte, which overcomes the problem of the repeated start.